+++ src/main/java/chatbot/client/domain/pingpong/PingPongServiceImpl.java
@@ -1,0 +1,20 @@
+package chatbot.client.domain.pingpong;
+
+import chatbot.client.core.message.MessageTemplate;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Service;
+
+import java.util.Optional;
+
+import static java.util.Optional.ofNullable;
+
+@Slf4j
+@Service
+public class PingPongServiceImpl implements PingPongService {
+    @Override
+    public Optional<MessageTemplate> makeResponse(MessageTemplate template, String content) {
+        log.info("content = {}", content);
+        template.convert("Pong!");
+        return ofNullable(template.isEmpty() ? null : template);
+    }
+}
